requirements.txt
python-telegram-bot==20.3
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
Pillow
PyPDF2==3.0.1
python-dotenv==1.0.0